Crew class.

R6 class for a crew.

Public fields

name

Character of length 1, crew name.

store

R6 store object.

worker_classes

Named list of R6ClassGenerator objects with worker definitions. The first one is the default.

workers

Named list of worker objects.

Methods

Public methods

class_crew$new()

class_crew$recruit()

class_crew$launch()

class_crew$sendable()

class_crew$send()

class_crew$receivable()

class_crew$receive()

class_crew$shutdown()

class_crew$dismiss()

class_crew$validate()

Method new()

Crew constructor.

Usage

class_crew$new(
  name = basename(tempfile(pattern = "crew_")),
  store = crew::class_store_local$new(),
  worker_classes = list(crew::class_worker_callr, crew::class_worker_future)
)

Arguments

name

Character of length 1, crew name.

store

R6 store object.

worker_classes

List of R6ClassGenerator objects (created with R6Class) with worker definitions and constructor.

max_workers

Positive integer, maximum number of workers in the crew.

Method recruit()

create worker objects from one or more worker definitions. Does not actually launch the new worker objects.

Usage

class_crew$recruit(workers = 1L, class = names(self$worker_classes)[[1]], ...)

Arguments

workers

Number of new worker objects to create.

class

Name of the worker class to use for launching workers. check the worker_classes field for possible names.

...

Named arguments to the worker constructor.

Method launch()

Launch all down workers in the crew.

Usage

class_crew$launch(tags = NULL)

Arguments

tags

Character vector of allowable tags of eligible workers.

Method sendable()

Determine if any worker is unassigned and ready to accept a new job.

Usage

class_crew$sendable(tags = NULL)

Arguments

tags

Character vector of allowable tags of eligible workers.

Method send()

Send a job to an available worker. Assumes sendable() on the crew returns TRUE.

Usage

class_crew$send(fun, args = list(), tags = NULL)

Arguments

fun

Function to run in the job.

args

Named list of arguments to fun()

tags

Character vector of allowable tags of eligible workers.

Method receivable()

Determine if any worker in the crew is done with its current job and the job output is available for collection.

Usage

class_crew$receivable(tags = NULL)

Arguments

tags

Character vector of allowable tags of eligible workers.

Method receive()

Choose a receivable worker and collect its job output.

Usage

class_crew$receive(tags = NULL)

Arguments

tags

Character vector of allowable tags of eligible workers.

Method shutdown()

Shut down one or more running workers.

Usage

class_crew$shutdown(
  workers = Inf,
  sendable_only = TRUE,
  up_only = TRUE,
  tags = NULL
)

Arguments

workers

Positive integer of length 1. Maximum number of workers to try to shut down. Does not count busy (non-sendable) workers.

sendable_only

Logical of length 1, whether to ignore workers that are not sendable.

up_only

Logical of length 1, whether to ignore workers that are not up.

tags

Character vector of allowable tags of eligible workers.

Details

This method loops through the workers from first to last in the worker list. If the worker is up (running) and sendable (able to receive a job) then the worker is sent a shutdown command. Otherwise, the loop moves on to another worker.

Method dismiss()

delete all worker objects.

Usage

class_crew$dismiss(
  workers = Inf,
  sendable_only = TRUE,
  down_only = TRUE,
  tags = NULL
)

Arguments

workers

Positive integer of length 1. Maximum number of workers to try to dismiss. Does not count up or busy (non-sendable) workers (workers with unfinished jobs).

sendable_only

Logical of length 1, whether to ignore workers that are not sendable.

down_only

Logical of length 1, whether to ignore workers that are still up.

tags

Character vector of allowable tags of eligible workers.

Details

This method loops through the workers from first to last in the worker list. If the worker is down and sendable (able to receive a job) then the worker object is deleted from the crew. Otherwise, the loop moves on to another worker.

Method validate()

Crew validator.

Usage

class_crew$validate()