Local store class.

R6 class for a local store.

Super class

crew::store -> store_local

Public fields

dir_temp

Character of length 1, directory for temporary files.

Methods

Public methods

class_store_local$new()

class_store_local$path_temp()

class_store_local$read_input()

class_store_local$read_output()

class_store_local$write_input()

class_store_local$write_output()

class_store_local$exists_input()

class_store_local$exists_output()

class_store_local$delete_input()

class_store_local$delete_output()

class_store_local$destroy()

class_store_local$marshal()

Inherited methods

crew::store$path_input()

crew::store$path_output()

crew::store$validate()

Method new()

Store constructor.

Usage

class_store_local$new(dir_root = tempfile())

Arguments

dir_root

Character of length 1, file path or prefix where all the files are located.

Method path_temp()

Path to a worker's temporary file.

Usage

class_store_local$path_temp(name)

Arguments

name

Worker name.

Method read_input()

Read worker input.

Usage

class_store_local$read_input(name)

Arguments

name

Character of length 1, Worker name.

Method read_output()

Read worker output.

Usage

class_store_local$read_output(name)

Arguments

name

Character of length 1, Worker name.

Method write_input()

Write worker input.

Usage

class_store_local$write_input(name, data)

Arguments

name

Character of length 1, Worker name.

data

Data to write.

Method write_output()

Write worker output.

Usage

class_store_local$write_output(name, data)

Arguments

name

Character of length 1, Worker name.

data

Data to write.

Method exists_input()

Exists worker input?

Usage

class_store_local$exists_input(name)

Arguments

name

Character of length 1, Worker name.

Method exists_output()

Exists worker output?

Usage

class_store_local$exists_output(name)

Arguments

name

Character of length 1, Worker name.

Method delete_input()

Delete worker input.

Usage

class_store_local$delete_input(name)

Arguments

name

Character of length 1, Worker name

Method delete_output()

Delete worker output.

Usage

class_store_local$delete_output(name)

Arguments

name

Character of length 1, Worker name

Method destroy()

Delete all the files in the data store.

Usage

class_store_local$destroy()

Arguments

name

Character of length 1, Worker name

Method marshal()

Marshal the R6 store object to ship to a worker.

Usage

class_store_local$marshal()

Returns

Character of length 1 with R code to run to reconstruct the store on a worker.