Worker class.

R6 class for a worker.

Public fields

name

Character of length 1, worker name.

crew

R6 crew object to which the worker belongs.

timeout

Positive numeric of length 1, number of seconds of idling for the worker to time out.

wait_input

Positive numeric of length 1, number of seconds for a worker to wait between iterations of polling input.

tags

Character vector of optional user-defined tags to select subsets of eligible workers for job submission and retrieval.

assigned

Logical of length 1, whether the worker already has a job to do.

Methods

Public methods

class_worker$new()

class_worker$sendable()

class_worker$send()

class_worker$receivable()

class_worker$receive()

class_worker$shutdown()

class_worker$tagged()

class_worker$validate()

Method new()

Worker constructor.

Usage

class_worker$new(
  name = basename(tempfile(pattern = "worker_")),
  crew = NULL,
  timeout = 60,
  wait_input = 0.1,
  tags = character(0)
)

Arguments

name

Character of length 1, worker name.

crew

R6 crew object to which the worker belongs.

timeout

Positive numeric of length 1, number of seconds that a worker can idle before timing out.

wait_input

Positive numeric of length 1, number of seconds that the worker waits between checking if a job exists.

tags

Character vector of optional user-defined tags to mark eligible groups of workers in scheduling operations.

Method sendable()

Check if this worker is ready to accept a job.

Usage

class_worker$sendable()

Method send()

Send a job.

Usage

class_worker$send(fun, args = list())

Arguments

fun

Function to run in the job. Should be completely self-contained in the body and arguments, without relying on the closure or global variables in the environment.

args

Named list of arguments to fun.

Method receivable()

TRUE if a worker is receivable with a job and the main process can receive the output of the job. FALSE otherwise.

Usage

class_worker$receivable()

Method receive()

Collect the results of a job.

Usage

class_worker$receive()

Method shutdown()

Gracefully shut down the worker.

Usage

class_worker$shutdown()

Method tagged()

Check if the worker has one or more of the tags in the argument.

Usage

class_worker$tagged(tags)

Arguments

tags

Character vector of tags to check.

Method validate()

Worker validator.

Usage

class_worker$validate()