callr worker class.

R6 class for a callr worker.

Super class

crew::worker -> worker_callr

Public fields

process

A callr::r_bg() process handler.

Methods

Public methods

class_worker_callr$launch()

class_worker_callr$up()

class_worker_callr$shutdown()

class_worker_callr$validate()

Inherited methods

crew::worker$initialize()

crew::worker$receivable()

crew::worker$receive()

crew::worker$send()

crew::worker$sendable()

crew::worker$tagged()

Method launch()

Launch the worker.

Usage

class_worker_callr$launch()

Returns

NULL (invisibly).

Method up()

Check if the underlying worker R process is actually running.

Usage

class_worker_callr$up()

Details

While running, the worker could be actually running a job, or it could be waiting for job input. This way, a worker can accept multiple jobs throughout its lifetime before it times out and is possibly relaunched. This is what it means to be a "semi-persistent" worker.

Returns

TRUE if the worker is running and FALSE otherwise.

Method shutdown()

Gracefully shut down the worker.

Usage

class_worker_callr$shutdown()

Details

The underlying worker process should promptly shut down if successful. A new send() or launch() call will re-launch the worker.

The shutdown method for a callr worker is quicker and more reliable than the default method of sending a shutdown job through the data store.

Returns

NULL (invisibly).

Method validate()

Worker validator.

Usage

class_worker_callr$validate()

Examples

crew <- class_crew$new(worker_classes = list(class_worker_callr))
crew$recruit(workers = 1)
worker <- crew$workers[[1]]
worker$send(fun = function(arg) paste("job", arg), args = list(arg = 1))
while (!worker$receivable()) Sys.sleep(0.1)
job <- worker$receive()
print(job$value)
#> [1] "job 1"
print(job$error)
#> NULL
worker$shutdown()