Future worker class.

R6 class for a future worker.

Super class

crew::worker -> worker_future

Public fields

future

A future::future() object.

Methods

Public methods

class_worker_future$launch()

class_worker_future$up()

class_worker_future$validate()

Inherited methods

crew::worker$clear()

crew::worker$initialize()

crew::worker$receivable()

crew::worker$receive()

crew::worker$restart()

crew::worker$send()

crew::worker$sendable()

crew::worker$shutdown()

crew::worker$stuck()

crew::worker$tagged()

Method launch()

Launch the worker.

Usage

class_worker_future$launch()

Returns

NULL (invisibly).

Method up()

TRUE if the worker is running and FALSE otherwise.

Usage

class_worker_future$up()

Details

While running, the worker could be actually running a job, or it could be waiting for job input. This way, a worker can accept multiple jobs throughout its lifetime before it times out and is possibly relaunched. This is what it means to be a "semi-persistent" worker.

Returns

TRUE if the worker is running and FALSE otherwise.

Method validate()

Worker validator.

Usage

class_worker_future$validate()

Examples

if (requireNamespace("future.callr", quietly = TRUE)) {
future::plan(future.callr::callr)
crew <- class_crew$new(worker_classes = list(class_worker_future))
crew$recruit(workers = 1)
worker <- crew$workers[[1]]
worker$send(fun = function(arg) paste("job", arg), args = list(arg = 1))
while (!worker$receivable()) Sys.sleep(0.1)
job <- worker$receive()
print(job$value)
print(job$error)
worker$shutdown()
future::plan(future::sequential)
processx::supervisor_kill()
}
#> [1] "job 1"
#> NULL
#> NULL