mirai launcher class

R6 class to launch and manage mirai workers.

Details

See crew_launcher_callr().

See also

Super class

crew::crew_class_launcher -> crew_class_launcher_callr

Methods

Public methods

crew_class_launcher_callr$launch_worker()

crew_class_launcher_callr$terminate_worker()

Inherited methods

crew::crew_class_launcher$active()

crew::crew_class_launcher$call()

crew::crew_class_launcher$clean()

crew::crew_class_launcher$inactive()

crew::crew_class_launcher$initialize()

crew::crew_class_launcher$launch()

crew::crew_class_launcher$populate()

crew::crew_class_launcher$settings()

crew::crew_class_launcher$terminate()

crew::crew_class_launcher$unreachable()

crew::crew_class_launcher$validate()

Method launch_worker()

Launch a callr worker to dial into a socket.

Usage

crew_class_launcher_callr$launch_worker(socket, host, port, token, name)

Arguments

socket

Socket where the worker will receive tasks.

host

IP address of the mirai client that sends tasks.

port

TCP port to register a successful connection to the host. Different from that of socket.

token

Character of length 1 to identify the instance of the mirai server process connected to the socket.

name

User-supplied name of the launcher, useful for constructing informative job labels.

Returns

A callr::r_bg() handle.

Method terminate_worker()

Terminate a callr worker.

Usage

crew_class_launcher_callr$terminate_worker(handle)

Arguments

handle

A callr process handle previously returned by launch_worker().

Returns

NULL (invisibly).

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
router <- crew_router()
router$listen()
launcher <- crew_launcher_callr()
launcher$populate(sockets = router$sockets)
launcher$launch()
m <- mirai::mirai("result")
Sys.sleep(0.25)
m$data
router$terminate()
crew_session_terminate()
}