mirai controller class

R6 class for mirai controllers.

Details

See crew_mirai_controller().

See also

Public fields

router

Router object.

launcher

Launcher object.

scale_method

Scaling method. See crew_mirai_controller().

queue

List of tasks in the queue.

results

List of finished tasks

Methods

Public methods

crew_class_mirai_controller$new()

crew_class_mirai_controller$validate()

crew_class_mirai_controller$connect()

crew_class_mirai_controller$launch()

crew_class_mirai_controller$collect()

crew_class_mirai_controller$scale()

crew_class_mirai_controller$push()

crew_class_mirai_controller$pop()

crew_class_mirai_controller$wait()

crew_class_mirai_controller$terminate()

crew_class_mirai_controller$clone()

Method new()

mirai controller constructor.

Usage

crew_class_mirai_controller$new(
  router = NULL,
  launcher = NULL,
  scale_method = NULL
)

Arguments

router

Router object. See crew_mirai_controller().

launcher

Launcher object. See crew_mirai_controller().

scale_method

Scaling method. See crew_mirai_controller(). #' @examples if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) router <- crew_mirai_router() # Use instead of the constructor. launcher <- crew_mirai_launcher_callr() controller <- crew_mirai_controller(router = router, launcher = launcher) controller$connect() controller$push(name = "task", command = sqrt(4)) controller$wait() controller$pop() controller$terminate()

Returns

An R6 object with the controller object.

Method validate()

Validate the router.

Usage

crew_class_mirai_controller$validate()

Returns

NULL (invisibly).

Method connect()

Connect the router and register the TCP sockets with the launcher.

Usage

crew_class_mirai_controller$connect()

Returns

NULL (invisibly).

Method launch()

Launch one or more workers.

Usage

crew_class_mirai_controller$launch(n = 1L)

Arguments

n

Maximum number of workers to launch.

Details

The actual number of newly launched workers is capped at the max set when the controller was created. Unless you are launching an array of persistent workers or have scale_method = "none", this method does not usually need to be called directly because push(scale = TRUE) already launches workers depending on scale_method.

Returns

NULL (invisibly).

Method collect()

Check for done tasks and move the results to the results list.

Usage

crew_class_mirai_controller$collect()

Returns

NULL (invisibly). Removes elements from the queue list as applicable and moves them to the results list.

Method scale()

Automatically scale up the number of workers if needed.

Usage

crew_class_mirai_controller$scale()

Details

This method is called during push(), and the method for scaling up workers is governed by the scale_method argument of crew_mirai_controller(). It is not meant to be called manually. If called manually, it is recommended to call collect() first so scale() can accurately assess the demand. For finer control of the number of workers launched, call launch() on the controller with the exact desired number of workers.

Returns

NULL (invisibly).

Method push()

Push a task to the head of the task list.

Usage

crew_class_mirai_controller$push(
  command,
  args = list(),
  timeout = NULL,
  name = NULL,
  scale = TRUE
)

Arguments

command

R code with the task to run.

args

A list of objects referenced in expr.

timeout

Time in milliseconds for the task.

name

Optional name of the task. Replaced with a random name if NULL or in conflict with an existing name in the task list.

scale

Logical, whether to automatically scale workers to meet demand. If TRUE, then collect() runs first so demand can be properly assessed before scaling.

Returns

NULL (invisibly).

Method pop()

Pop a completed task from the results data frame.

Usage

crew_class_mirai_controller$pop(collect = TRUE)

Arguments

collect

Whether to run the collect() method to collect all available results before calling pop().

Returns

If there is a completed task available to collect, the return value is a one-row data frame with the results, warnings, and errors. Otherwise, if there are no results available to collect, the return value is NULL.

Method wait()

Wait for a result to be ready.

Usage

crew_class_mirai_controller$wait(timeout = Inf, wait = 0.1)

Arguments

timeout

Timeout length in seconds waiting for results to become available.

wait

Number of seconds to wait between polling intervals while checking for results.

Returns

NULL (invisibly). Call pop() to get the result.

Method terminate()

Terminate the workers and disconnect the router.

Usage

crew_class_mirai_controller$terminate()

Returns

NULL (invisibly).

Method clone()

The objects of this class are cloneable with this method.

Usage

crew_class_mirai_controller$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
launcher <- crew_mirai_launcher_callr()
controller <- crew_mirai_controller(router = router, launcher = launcher)
controller$connect()
controller$push(name = "task", command = sqrt(4))
controller$wait()
controller$pop()
controller$terminate()
}