mirai launcher class

R6 class to launch and manage mirai workers.

Details

See crew_mirai_launcher_callr().

See also

Public fields

sockets

TCP sockets for listening to the workers.

workers

List of callr::r_bg() handles for workers. The handle is NA if it is not yet called.

idle_time

Argument to crew_mirai_launcher_callr().

wall_time

Argument to crew_mirai_launcher_callr().

poll_high

Argument to crew_mirai_launcher_callr().

poll_low

Argument to crew_mirai_launcher_callr().

launch_timeout

Argument to crew_mirai_launcher_callr().

launch_wait

Argument to crew_mirai_launcher_callr().

max_tasks

Argument to crew_mirai_launcher_callr().

async_dial

Argument to crew_mirai_launcher_callr().

Methods

Public methods

crew_class_mirai_launcher_callr$new()

crew_class_mirai_launcher_callr$validate()

crew_class_mirai_launcher_callr$populate()

crew_class_mirai_launcher_callr$running()

crew_class_mirai_launcher_callr$launch()

crew_class_mirai_launcher_callr$terminate()

Method new()

mirai launcher constructor.

Usage

crew_class_mirai_launcher_callr$new(
  sockets = NULL,
  workers = NULL,
  idle_time = NULL,
  wall_time = NULL,
  poll_high = NULL,
  poll_low = NULL,
  launch_timeout = NULL,
  launch_wait = NULL,
  max_tasks = NULL,
  async_dial = NULL
)

Arguments

sockets

Argument to crew_mirai_launcher_callr().

workers

Argument to crew_mirai_launcher_callr().

idle_time

Argument to crew_mirai_launcher_callr().

wall_time

Argument to crew_mirai_launcher_callr().

poll_high

Argument to crew_mirai_launcher_callr().

poll_low

Argument to crew_mirai_launcher_callr().

launch_timeout

Argument to crew_mirai_launcher_callr().

launch_wait

Argument to crew_mirai_launcher_callr().

max_tasks

Argument to crew_mirai_launcher_callr().

async_dial

Argument to crew_mirai_launcher_callr().

Returns

An R6 object with the launcher.

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$connect()
launcher <- crew_mirai_launcher_callr()
launcher$populate(sockets = router$sockets_listening())
launcher$running() # 0
launcher$launch()
launcher$running() # 1
launcher$terminate()
launcher$running() # 0
router$disconnect()
}

Method validate()

Validate the launcher.

Usage

crew_class_mirai_launcher_callr$validate()

Returns

NULL (invisibly).

Method populate()

Populate workers.

Usage

crew_class_mirai_launcher_callr$populate(sockets = character(0))

Arguments

sockets

Character vector of local TCP sockets that the workers will use to dial in to receive tasks.

Returns

NULL (invisibly).

Method running()

Count the number of running workers.

Usage

crew_class_mirai_launcher_callr$running()

Returns

Number of running workers.

Method launch()

Launch one or more workers.

Usage

crew_class_mirai_launcher_callr$launch(n = 1L)

Arguments

n

Maximum number of workers to launch.

Details

The actual number of newly launched workers is less than or equal to the number of sockets supplied to $populate() minus the number of workers already running.

Returns

NULL (invisibly).

Method terminate()

Terminate all workers.

Usage

crew_class_mirai_launcher_callr$terminate()

Returns

NULL (invisibly).

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$connect()
launcher <- crew_mirai_launcher_callr()
launcher$populate(sockets = router$sockets_listening())
launcher$running() # 0
launcher$launch()
launcher$running() # 1
launcher$terminate()
launcher$running() # 0
router$disconnect()
}

## ------------------------------------------------
## Method `crew_class_mirai_launcher_callr$new`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$connect()
launcher <- crew_mirai_launcher_callr()
launcher$populate(sockets = router$sockets_listening())
launcher$running() # 0
launcher$launch()
launcher$running() # 1
launcher$terminate()
launcher$running() # 0
router$disconnect()
}