Router class

R6 class for mirai routers.

Details

See crew_router().

See also

Public fields

name

See crew_router().

workers

See crew_router().

host

See crew_router().

port

See crew_router().

seconds_interval

See crew_router().

seconds_timeout

See crew_router().

started

Whether the router is started.

dispatcher

Process ID of the mirai dispatcher

daemons

Data frame of information from mirai::daemons().

rotations

Logical vector to keep track of rotated worker websockets.

Methods

Public methods

crew_class_router$new()

crew_class_router$validate()

crew_class_router$start()

crew_class_router$route()

crew_class_router$poll()

crew_class_router$log()

crew_class_router$terminate()

Method new()

mirai router constructor.

Usage

crew_class_router$new(
  name = NULL,
  workers = NULL,
  host = NULL,
  port = NULL,
  seconds_interval = NULL,
  seconds_timeout = NULL
)

Arguments

name

Argument passed from crew_router().

workers

Argument passed from crew_router().

host

Argument passed from crew_router().

port

Argument passed from crew_router().

seconds_interval

Argument passed from crew_router().

seconds_timeout

Argument passed from crew_router().

Returns

An R6 object with the router.

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$start()
router$daemons
router$terminate()
}

Method validate()

Validate the router.

Usage

crew_class_router$validate()

Returns

NULL (invisibly).

Method start()

Start listening for workers on the available sockets.

Usage

crew_class_router$start()

Returns

NULL (invisibly).

Method route()

Choose the websocket path for the next instance of the worker at a given index.

Usage

crew_class_router$route(index, force = FALSE)

Arguments

index

Integer of length 1, worker index.

force

TRUE to force a rotation even if e.g. tasks are underway, FALSE otherwise.

Details

The first call to route() at a given index uses the original websocket path provided by mirai::daemons(). Subsequent calls to route() at the same index rotate the websocket path for robustness.

Returns

Character of length 1, new websocket path of the worker.

Method poll()

Update the daemons field with information on the mirai daemons.

Usage

crew_class_router$poll()

Details

Call mirai::daemons() to get information about the workers. If the workers cannot be reached or the router has not started, then do not modify the daemons field. Otherwise, if the workers are reachable, populate the daemons field with a matrix of high-level worker-specific statistics.

Returns

NULL (invisibly).

Method log()

Show an informative worker log.

Usage

crew_class_router$log()

Details

This method tries once to update the worker information using poll(). If unsuccessful the first time, it just uses the existing information in the daemons field.

Returns

A tibble with information on the workers.

Method terminate()

Stop the mirai client and disconnect from the worker websockets.

Usage

crew_class_router$terminate()

Returns

NULL (invisibly).

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$start()
router$daemons
router$terminate()
}

## ------------------------------------------------
## Method `crew_class_router$new`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$start()
router$daemons
router$terminate()
}