Create a client object.

Create an R6 wrapper object to manage the mirai client.

Usage

crew_client(
  name = NULL,
  workers = 1L,
  host = NULL,
  port = NULL,
  tls = crew::crew_tls(),
  tls_enable = NULL,
  tls_config = NULL,
  seconds_interval = 0.5,
  seconds_timeout = 5,
  retry_tasks = TRUE,
  log_resources = NULL
)

Arguments

name

Name of the client object. If NULL, a name is automatically generated.

workers

Integer, maximum number of parallel workers to run.

host

IP address of the mirai client to send and receive tasks. If NULL, the host defaults to the local IP address.

port

TCP port to listen for the workers. If NULL, then an available ephemeral port is automatically chosen.

tls

A TLS configuration object from crew_tls().

tls_enable

Deprecated on 2023-09-15 in version 0.4.1. Use argument tls instead.

tls_config

Deprecated on 2023-09-15 in version 0.4.1. Use argument tls instead.

seconds_interval

Number of seconds between polling intervals waiting for certain internal synchronous operations to complete, such as checking mirai::status()

seconds_timeout

Number of seconds until timing out while waiting for certain synchronous operations to complete, such as checking mirai::status().

retry_tasks

TRUE to automatically retry a task in the event of an unexpected worker exit. FALSE to give up on the first exit and return a mirai error code (code number 19). TRUE (default) is recommended in most situations. Use FALSE for debugging purposes, e.g. to confirm that a task is causing a worker to run out of memory or crash in some other way.

log_resources

Optional character string with a file path to a text file to log memory consumption. Set log_resources to NULL to avoid writing to a log file. If you supply a path, then the log() method will write memory usage statistics to the file, and most controller methods will do the same with throttling so resource consumption is recorded throughout the whole life cycle of the controller.

The log file is in comma-separated values (CSV) format which can be easily read by readr::read_csv(). The controller automatically deletes the old log file when it starts (when controller$start() is called for the first time, but not subsequent times).

The log file has one row per observation of a process, including the current R process ("client") and the mirai dispatcher. If the dispatcher is not included in the output, it means the dispatcher process is not running. Columns include: * type: the type of process (client or dispatcher) * pid: the process ID. * status: The process status (from ps::ps_status()). * rss: resident set size (RSS). RS is the total memory held by a process, including shared libraries which may also be in use by other processes. RSS is obtained from ps::ps_memory_info() and shown in bytes. * elapsed: number of elapsed seconds since the R process was started (from proc.time()["elapsed"]).

See also

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
client <- crew_client()
client$start()
client$summary()
client$terminate()
}