Create a mirai controller with a callr launcher.

Create an R6 object to submit tasks and launch workers with a mirai task queue and callr workers

crew_mirai_controller_callr(
  name = NULL,
  workers = 1L,
  host = NULL,
  ports = NULL,
  router_timeout = 5,
  router_wait = 0.1,
  idle_time = Inf,
  wall_time = Inf,
  poll_high = 5,
  poll_low = 50,
  launch_timeout = 5,
  launch_wait = 0.1,
  max_tasks = Inf,
  async_dial = TRUE,
  scale_method = "demand"
)

Arguments

name

Name of the mirai router. Defaults to a string from ids::proquint().

workers

Integer, maximum number of parallel workers to run. crew will reserve one ephemeral port for each worker. See the Ports section for an important cautionary note.

host

IP address of the client process that the workers can dial into inside the local network. If a character string, the router uses the specified IP address. If NULL, the IP address defaults to getip::getip(type = "local").

ports

Optional integer vector of TCP port numbers. Supersedes workers if supplied. Except for 0, which defers to NNG to automatically assign a port, each unique port value corresponds to a local port where a worker will dial in to accept tasks.

router_timeout

Number of seconds to time out waiting for the mirai router to (dis)connect.

router_wait

Number of seconds to wait between iterations checking if the mirai router is (dis)connected.

idle_time

Maximum number of seconds that a worker can idle since the completion of the last task. If exceeded, the worker exits.

wall_time

Soft wall time in seconds. See the wall_time argument of mirai::server().

poll_high

High polling interval in seconds for the mirai active queue.

poll_low

Low polling interval in seconds for the mirai active queue.

launch_timeout

Number of seconds to time out waiting for a new group of workers to launch.

launch_wait

Number of seconds to wait between checks that newly launched workers are ready to receive tasks.

max_tasks

Maximum number of tasks that a worker will do before exiting.

async_dial

Logical, whether the mirai workers should dial in asynchronously. See the asyncdial argument of mirai::server().

scale_method

Character of length 1, name of the method for automatically scaling workers to meet demand. NULL to default to "demand". Possible values include the following:

"demand": just after pushing a new task in push(), launch min(n, max(0, t - w)) workers, where n is the maximum number of workers, t is the number of queued tasks, and w is the current number of workers already running. In other words, scale up the number of workers to meet the current demand.

"single": just after pushing a new task in push(), launch a single worker if demand min(n, max(0, t - w)) is greater than 0.

"none": do not auto-scale at all.

See also

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
controller <- crew_mirai_controller_callr()
controller$connect()
controller$push(name = "task", command = sqrt(4))
controller$wait()
controller$pop()
controller$terminate()
}