Create a mirai launcher with callr workers.

Create an R6 object to launch and maintain callr workers for a mirai controller.

crew_mirai_launcher_callr(
  idle_time = Inf,
  wall_time = Inf,
  poll_high = 5,
  poll_low = 50,
  launch_timeout = 5,
  launch_wait = 0.1,
  max_tasks = Inf,
  async_dial = TRUE
)

Arguments

idle_time

Maximum number of seconds that a worker can idle since the completion of the last task. If exceeded, the worker exits.

wall_time

Soft wall time in seconds. See the wall_time argument of mirai::server().

poll_high

High polling interval in seconds for the mirai active queue.

poll_low

Low polling interval in seconds for the mirai active queue.

launch_timeout

Number of seconds to time out waiting for a new group of workers to launch.

launch_wait

Number of seconds to wait between checks that newly launched workers are ready to receive tasks.

max_tasks

Maximum number of tasks that a worker will do before exiting.

async_dial

Logical, whether the mirai workers should dial in asynchronously. See the asyncdial argument of mirai::server().

See also

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$connect()
launcher <- crew_mirai_launcher_callr()
launcher$populate(sockets = router$sockets_listening())
launcher$running() # 0
launcher$launch()
launcher$running() # 1
launcher$terminate()
launcher$running() # 0
router$disconnect()
}