Create a mirai router.

Create an R6 object to route tasks to mirai workers on the local network.

crew_mirai_router(
  name = NULL,
  workers = 1L,
  host = NULL,
  ports = NULL,
  router_timeout = 5,
  router_wait = 0.1
)

Arguments

name

Name of the mirai router. Defaults to a string from ids::proquint().

workers

Integer, maximum number of parallel workers to run. crew will reserve one ephemeral port for each worker. See the Ports section for an important cautionary note.

host

IP address of the client process that the workers can dial into inside the local network. If a character string, the router uses the specified IP address. If NULL, the IP address defaults to getip::getip(type = "local").

ports

Optional integer vector of TCP port numbers. Supersedes workers if supplied. Except for 0, which defers to NNG to automatically assign a port, each unique port value corresponds to a local port where a worker will dial in to accept tasks.

router_timeout

Number of seconds to time out waiting for the mirai router to (dis)connect.

router_wait

Number of seconds to wait between iterations checking if the mirai router is (dis)connected.

Ports

In the mirai-based task scheduling in crew, each parallel worker dials into a different TCP port on the local machine. If you launch hundreds of workers, then hundreds of ports will not be available to other users or processes. Large numbers of workers on shared machines or clusters may seriously disrupt the tasks of other users, so please be careful.

See also

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_mirai_router()
router$sockets_listening() # character(0)
router$connect()
router$sockets_listening() # "tcp://xx.xx.xx:xxxxx"
router$disconnect()
router$sockets_listening() # character(0)
}