Abstract queue

Abstract class for a task queue.

Details

This is an abstract class and not runnable by itself. Please instantiate using a subclass such as crew_queue_session.

Queue attribution

The design and implementation of crew task queues, especially the local callr-based task queues, borrows heavily from the work of Gábor Csárdi, particularly https://www.tidyverse.org/blog/2019/09/callr-task-q/, which was also released in commit 811a02f604de2cf03264f6b35ce9ec8a412f2581 of https://github.com/r-lib/callr in the vignettes/Task-queue-with-callr.Rmd file under the MIT license. See the crew package README.md and NOTICE files for other mentions of this attribution.

See also

Methods

Public methods

crew_queue$new()

crew_queue$get_tasks()

crew_queue$get_results()

crew_queue$get_workers()

crew_queue$push()

crew_queue$pop()

crew_queue$update()

crew_queue$crashed()

crew_queue$block()

crew_queue$shutdown()

Method new()

Abstract queue constructor.

Usage

crew_queue$new(
  workers = 1,
  store = crew_store_local$new(timeout = timeout, wait = wait),
  timeout = 60,
  wait = 0.1,
  max_tasks = Inf
)

Arguments

workers

Number of workers in the queue.

store

Data store object created with crew_store_local or similar.

timeout

Number of seconds to for a worker to wait for something to happen (e.g. the arrival of a task) before timing out and quitting.

wait

Number of seconds to wait in between iterations while waiting for something to happen (e.g. the arrival of a task).

max_tasks

Number of tasks a worker can run before quitting. Some queues allow the worker to restart when launching a new task.

Returns

An abstract queue object.

Method get_tasks()

Show the queued tasks pushed with the push() method but not yet assigned to workers.

Usage

crew_queue$get_tasks()

Returns

A data frame with one row per queued task and columns describing each task.

Method get_results()

Show the results completed by the workers and available to retrieve with the pop() method.

Usage

crew_queue$get_results()

Returns

A data frame with one row per completed task and columns describing each task.

Method get_workers()

Show the workers of the queue.

Usage

crew_queue$get_workers()

Returns

A data frame with one row per worker and columns describing each worker.

Method push()

Push a new task on to the queue.

Usage

crew_queue$push(fun, args = list(), task = crew_name(), update = TRUE, ...)

Arguments

fun

R function that runs the task.

args

Named list of arguments to fun.

task

Character of length 1 with the task ID.

update

Logical of length 1, whether to update the internal state of the queue after pushing. See the update() method for details.

...

Named arguments specific to each job, such as the future plan for crew_queue_future queues.

Returns

NULL (invisibly)

Method pop()

Pop the results of a completed task off the queue.

Usage

crew_queue$pop(update = TRUE)

Arguments

update

Logical of length 1, whether to update the internal state of the queue after pushing. See the update() method for details.

Returns

A named list with the results of the task.

Method update()

Update the internal state of the queue.

Usage

crew_queue$update()

Details

Updating moves tasks along and make sure they get assigned to workers and cleaned up in a timely manner. However, there may be a computational overhead cost if updates happen too often. It is recommended to have an event loop that updates the queue, (preferably via methods push() and pop() instead of update() directly) that updates the queue at a time interval that appropriately balances the tradeoffs.

Returns

NULL (invisibly)

Method crashed()

Detect crashed workers.

Usage

crew_queue$crashed()

Details

This usually happens during update() anyway.

Returns

Return NULL (invisibly) if no workers crashed. Throw an error if at least one worker crashed.

Method block()

Block the R session until there are no unassigned tasks and at least one worker is available.

Usage

crew_queue$block(timeout = NULL, wait = NULL)

Arguments

timeout

Number of seconds to wait to clear the queue backlog. Defaults to the timeout value assigned at initialization.

wait

Number of seconds to wait in between update iterations. Defaults to the wait value assigned at initialization.

Details

The block() method calls update() every wait seconds until the backlog is clear or timeout is reached. It is a convenient event loop if you want to clear the backlog before taking any other action, but most event loops continuously assign tasks with push() and retrieve available ones with pop().

Returns

NULL (invisibly)

Method shutdown()

Shut down the workers.

Usage

crew_queue$shutdown()

Details

Not all queues can shut down all their workers. For example, shutdown() in crew_queue_future can shut down its callr-based subqueue, but it cannot shut down its future workers. Local subqueues such as crew_queue_session and crew_queue_bg can shut down completely.

Returns

NULL (invisibly)

Examples

fun <- function(x) x + 1
args <- list(x = 1)
queue <- crew_queue_session$new(timeout = 60, wait = 0.1)
queue$push(fun = fun, args = args)
queue$block()
result <- queue$pop()
str(result)
#> List of 2
#>  $ task  : chr "xdb0e5afe_b1dc_481f_9a7d_7b58a1347602"
#>  $ result:List of 6
#>   ..$ code   : int 200
#>   ..$ message: chr "done callr-rs-result-35324ebf51a8"
#>   ..$ result : num 2
#>   ..$ stdout : chr ""
#>   ..$ stderr : chr ""
#>   ..$ error  : NULL
#>   ..- attr(*, "class")= chr "callr_session_result"
result$result$result
#> [1] 2
queue$shutdown()
processx::supervisor_kill()