Local background queue

Task queue with callr::r_bg() workers.

Details

Some public methods are inherited from crew_queue. See the crew_queue help file for details on those.

The interface is the same as crew_queue_session except there is a data store (e.g. crew_store_local) which the user can customize. Most users should choose for the session queue crew_queue_session instead of the background queue crew_queue_bg. The only reason to use crew_queue_bg is for subqueues for crew_queue_future on clusters where batchtools futures clash with callr::r_session.

Queue attribution

The design and implementation of crew task queues, especially the local callr-based task queues, borrows heavily from the work of Gábor Csárdi, particularly https://www.tidyverse.org/blog/2019/09/callr-task-q/, which was also released in commit 811a02f604de2cf03264f6b35ce9ec8a412f2581 of https://github.com/r-lib/callr in the vignettes/Task-queue-with-callr.Rmd file under the MIT license. See the crew package README.md and NOTICE files for other mentions of this attribution.

See also

Super class

crew::crew_queue -> crew_queue_bg

Methods

Public methods

crew_queue_bg$new()

crew_queue_bg$shutdown()

Inherited methods

crew::crew_queue$block()

crew::crew_queue$crashed()

crew::crew_queue$get_results()

crew::crew_queue$get_tasks()

crew::crew_queue$get_workers()

crew::crew_queue$pop()

crew::crew_queue$push()

crew::crew_queue$update()

Method new()

Local background queue constructor.

Usage

crew_queue_bg$new(
  workers = 1,
  timeout = 60,
  wait = 0.1,
  store = crew_store_local$new(timeout = Inf, wait = wait)
)

Arguments

workers

Number of workers in the queue.

timeout

Number of seconds to for a worker to wait for something to happen (e.g. the arrival of a task) before timing out and quitting.

wait

Number of seconds to wait in between iterations while waiting for something to happen (e.g. the arrival of a task).

store

A local store object from crew_store_local.

Returns

A local background queue object.

Method shutdown()

Shut down the workers.

Usage

crew_queue_bg$shutdown()

Returns

NULL (invisibly)

Examples

fun <- function(x) x + 1
args <- list(x = 1)
queue <- crew_queue_bg$new(timeout = 60, wait = 0.1)
queue$push(fun = fun, args = args)
queue$block()
result <- queue$pop()
str(result)
#> List of 2
#>  $ task  : chr "x63894af5_51bc_4062_84a1_f112d7f475c1"
#>  $ result:List of 5
#>   ..$ result   : num 2
#>   ..$ seconds  : num 0
#>   ..$ error    : NULL
#>   ..$ traceback: NULL
#>   ..$ warnings : NULL
result$result$result
#> [1] 2
queue$shutdown()
processx::supervisor_kill()