Future queue

Task queue with future workers.

Details

The workers in this queue are futures with the future package. In push(), it is possible to supply different future::plan()s to different futures, and the queue itself has its own default plan. Plans must be serializable, so only the sequential and future.batchtools plans are supported. The queue has a subqueue from crew_queue_session (or crew_queue_bg if absolutely necessary) to submit and poll futures in order to eliminate the otherwise heavy overhead of future::future() and future::resolved() in future.batchtools futures. Tasks are sent and collected using a crew_store_local data store in order to eliminate the normally heavy overhead of future::value().

Queue attribution

The design and implementation of crew task queues, especially the local callr-based task queues, borrows heavily from the work of Gábor Csárdi, particularly https://www.tidyverse.org/blog/2019/09/callr-task-q/, which was also released in commit 811a02f604de2cf03264f6b35ce9ec8a412f2581 of https://github.com/r-lib/callr in the vignettes/taskq.R file under the MIT license. See the crew package README.md and NOTICE files for other mentions of this attribution.

See also

Super class

crew::crew_queue -> crew_queue_future

Methods

Public methods

crew_queue_future$new()

crew_queue_future$get_plan()

crew_queue_future$push()

Inherited methods

crew::crew_queue$block()

crew::crew_queue$crashed()

crew::crew_queue$get_results()

crew::crew_queue$get_tasks()

crew::crew_queue$get_workers()

crew::crew_queue$pop()

crew::crew_queue$shutdown()

crew::crew_queue$update()

Method new()

Abstract queue constructor.

Usage

crew_queue_future$new(
  workers = 1,
  store = crew_store_local$new(timeout = timeout, wait = wait),
  timeout = 60,
  wait = 0.1,
  max_tasks = Inf,
  plan = future::plan(),
  subqueue = crew_queue_session$new(workers = 1, wait = wait, timeout = timeout)
)

Arguments

workers

Number of workers in the queue.

store

Data store object created with crew_store_local or similar.

timeout

Number of seconds to for a worker to wait for something to happen (e.g. the arrival of a task) before timing out and quitting.

wait

Number of seconds to wait in between iterations while waiting for something to happen (e.g. the arrival of a task).

max_tasks

Number of tasks a worker can run before quitting. Some queues allow the worker to restart when launching a new task.

plan

A future::plan() object with the default plan.

subqueue

An object from crew_queue_session (or crew_queue_bg if absolutely necessary) with the subqueue which submits and collects futures in order to reduce overhead.

Returns

An abstract queue object.

Method get_plan()

Get the default future plan of the queue.

Usage

crew_queue_future$get_plan()

Returns

The default future plan object of the queue.

Method push()

Push a new task on to the queue.

Usage

crew_queue_future$push(
  fun,
  args = list(),
  task = crew_name(),
  update = TRUE,
  plan = NULL
)

Arguments

fun

R function that runs the task.

args

Named list of arguments to fun.

task

Character of length 1 with the task ID. This ID gets passed to the label object of future::future() which shows up on htop and cluster viewing commands such as squeue for SLURM and qstat for SGE.

update

Logical of length 1, whether to update the internal state of the queue after pushing. See the update() method for details.

plan

An optional future plan object for the task.

Returns

NULL (invisibly)

Examples

fun <- function(x) x + 1
args <- list(x = 1)
queue <- crew_queue_future$new(timeout = 60, wait = 0.1)
queue$push(fun = fun, args = args)
queue$block()
result <- queue$pop()
str(result)
#> List of 2
#>  $ task  : chr "x3a4747de_3527_478d_b989_be5b90287e5b"
#>  $ result:List of 5
#>   ..$ result   : num 2
#>   ..$ seconds  : num 0
#>   ..$ error    : NULL
#>   ..$ traceback: NULL
#>   ..$ warnings : NULL
result$result$result
#> [1] 2
queue$shutdown()
processx::supervisor_kill()