Random name

Generate a random string that can be used as a name for a worker or task.

Usage

crew_random_name(n = 12L)

Arguments

n

Number of bytes of information in the random string hashed to generate the name. Larger n is more likely to generate unique names, but it may be slower to compute.

Value

A random character string.

Details

The randomness is not reproducible and cannot be set with e.g. set.seed() in R.

See also

Examples

crew_random_name()
#> [1] "46d56a81d53b63b7711b2314"