Create a router.

Create an R6 object to manage the mirai task scheduler client.

crew_router(
  name = NULL,
  workers = 1L,
  host = NULL,
  port = NULL,
  seconds_interval = 0.01,
  seconds_timeout = 5,
  pad_daemons = FALSE
)

Arguments

name

Name of the router object. If NULL, a name is automatically generated.

workers

Integer, maximum number of parallel workers to run.

host

IP address of the mirai client to send and receive tasks. If NULL, the host defaults to the local IP address.

port

TCP port to listen for the workers. If NULL, then an available ephemeral port is automatically chosen.

seconds_interval

Number of seconds between polling intervals waiting for certain internal synchronous operations to complete. If space_poll is TRUE, then this is also the minimum number of seconds between calls to mirai::daemons() for the purposes of checking worker status.

seconds_timeout

Number of seconds until timing out while waiting for certain synchronous operations to complete.

pad_daemons

Logical of length 1, whether to enforce a minimum seconds_interval time interval between calls to mirai::daemons() to check worker status. Set to FALSE to boost efficiency on robust systems or TRUE to safeguard against issues on finicky systems. If the most recent call to mirai::daemons() is too soon, the router object will block with nanonext::msleep() for the remainder of the interval.

See also

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
router <- crew_router()
router$start()
router$daemons
router$terminate()
}