Abstract store

Abstract class for a data store.

Details

Most crew queues use data stores to send and receive task data. Some use entirely local data, others may use cloud data to support cloud workers. Stores have a "main" file space and a "worker" file space. The main file space is usually on the local machine, and the worker file space may be in the cloud. Sending and collecting data may involve shuffling files to and and the worker file space.

See also

Methods

Public methods

crew_store$new()

crew_store$get_root()

crew_store$read_input()

crew_store$read_output()

crew_store$write_input()

crew_store$write_output()

crew_store$exists_input()

crew_store$exists_output()

crew_store$list_input()

crew_store$list_output()

crew_store$delete_input()

crew_store$delete_output()

crew_store$path_log()

crew_store$marshal()

crew_store$destroy()

Method new()

Abstract store constructor.

Usage

crew_store$new(root = tempfile(), timeout = 60, wait = 1)

Arguments

root

Character of length 1, root directory of the data store on the local file system.

timeout

Numeric of length 1, number of seconds to wait to receive worker input or shuffle files in the worker loop.

wait

Numeric of length 1, number of seconds to pause between iterations waiting for files to arrive in their proper places in internal operations.

Returns

An abstract store object.

Method get_root()

Get the root directory of the data store.

Usage

crew_store$get_root()

Returns

The root directory of the data store.

Method read_input()

Read worker input data.

Usage

crew_store$read_input(worker)

Arguments

worker

Character of length 1, name of the worker.

Returns

Worker input data.

Method read_output()

Read worker output data.

Usage

crew_store$read_output(worker)

Arguments

worker

Character of length 1, name of the worker.

Returns

Worker output data.

Method write_input()

Write worker input data to the main file space.

Usage

crew_store$write_input(worker, value)

Arguments

worker

Character of length 1, name of the worker.

value

R object to write.

Returns

NULL (invisibly).

Method write_output()

Write worker output data to the main file space.

Usage

crew_store$write_output(worker, value)

Arguments

worker

Character of length 1, name of the worker.

value

R object to output.

Returns

NULL (invisibly).

Method exists_input()

Check if worker input data exists.

Usage

crew_store$exists_input(worker)

Arguments

worker

Character of length 1, name of the worker.

Returns

Logical of length 1, whether worker input data exists.

Method exists_output()

Check if worker output data exists.

Usage

crew_store$exists_output(worker)

Arguments

worker

Character of length 1, name of the worker.

Returns

Logical of length 1, whether worker output data exists.

Method list_input()

List workers with input data.

Usage

crew_store$list_input()

Returns

Character vector of workers with input data.

Method list_output()

List workers with output data.

Usage

crew_store$list_output()

Returns

Character vector of workers with output data.

Method delete_input()

Delete worker input data.

Usage

crew_store$delete_input(worker)

Arguments

worker

Character of length 1, name of the worker.

Returns

NULL (invisibly).

Method delete_output()

Delete worker output data.

Usage

crew_store$delete_output(worker)

Arguments

worker

Character of length 1, name of the worker.

Returns

NULL (invisibly).

Method path_log()

Get worker log path. Not valid for all stores.

Usage

crew_store$path_log(worker)

Arguments

worker

Character of length 1, name of the worker.

Returns

Character of length 1, path to the worker log file.

Method marshal()

Marshal the data store.

Usage

crew_store$marshal()

Details

Represent the store object as a concise character string that can be recovered with parse() and eval() in a different R process.

Returns

A concise character string that returns the data store object when evaluated with parse() and eval() in a different R process.

Method destroy()

Delete the files of the data store.

Usage

crew_store$destroy()

Details

The local files, along with the main file space, are always deleted. Other deletions may depend on the specific store subclass.

Returns

NULL (invisibly).

Examples

# Usage on its own:
store <- crew_store_local$new(timeout = Inf)
store$write_input("worker", list(value = "x"))
store$list_input()
#> [1] "worker"
store$exists_input("worker")
#> [1] TRUE
store$read_input("worker")
#> $value
#> [1] "x"
#> 
# Usage in a queue:
fun <- function(x) x + 1
args <- list(x = 1)
queue <- crew_queue_future$new(timeout = 60, store = store)
queue$push(fun = fun, args = args)
queue$block()
result <- queue$pop()
str(result)
#> List of 2
#>  $ task  : chr "xdfe7f12f_8751_492b_b26e_a1774d35ce6f"
#>  $ result:List of 5
#>   ..$ result   : num 2
#>   ..$ seconds  : num 0
#>   ..$ error    : NULL
#>   ..$ traceback: NULL
#>   ..$ warnings : NULL
result$result$result
#> [1] 2
queue$shutdown()
processx::supervisor_kill()
store$destroy()