Run a task in a crew worker.

Not a user-side function. Do not invoke directly.

crew_task(worker, store, timeout, wait)

Arguments

worker

Character of length 1, name of the worker.

store

Either a data store object or a character string with a marshaled data store.

timeout

Positive numeric of length 1, number of seconds that a worker can idle before timing out.

wait

Positive numeric of length 1, number of seconds that the worker waits between checking if a task exists.

Examples

if (!identical(Sys.getenv("CREW_EXAMPLES", unset = ""), "")) {
root <- tempfile()
dir.create(root)
store <- crew_store_local$new(root = root)
fun <- function(x) {
  x + 1
}
args <- list(x = 1)
value <- list(fun = deparse(fun), args = args, task = "my_task")
store$write_input(worker = "my_worker", value = value)
crew_task(
  worker = "my_worker",
  store = store$marshal(),
  timeout = 0,
  wait = 0
)
store$read_output("my_worker")$result # 2
}