Crew class.

R6 class for a crew.

Details

A crew object is an interface to manage multiple high-performance computing workers. Supported methods send jobs, receive output, and poll, etc.

Public fields

name

Character of length 1, crew name.

store

R6 store object.

worker_classes

Named list of R6ClassGenerator objects with worker definitions. The first one is the default.

workers

Named list of worker objects.

Methods

Public methods

class_crew$new()

class_crew$recruit()

class_crew$launch()

class_crew$sendable()

class_crew$send()

class_crew$receivable()

class_crew$receive()

class_crew$clear()

class_crew$shutdown()

class_crew$dismiss()

class_crew$restart()

class_crew$validate()

Method new()

Crew constructor.

Usage

class_crew$new(
  name = basename(tempfile(pattern = "crew_")),
  store = crew::class_store_local$new(),
  worker_classes = list(crew::class_worker_callr, crew::class_worker_future)
)

Arguments

name

Character of length 1, crew name.

store

R6 store object.

worker_classes

List of R6ClassGenerator objects (created with R6Class) with worker definitions and constructor.

max_workers

Positive integer, maximum number of workers in the crew.

Returns

The new() method calls the constructor and returns a new crew object.

Method recruit()

create worker objects from one or more worker definitions. Does not actually launch the new worker objects.

Usage

class_crew$recruit(workers = 1L, class = names(self$worker_classes)[[1]], ...)

Arguments

workers

Number of new worker objects to create.

class

Name of the worker class to use for launching workers. check the worker_classes field for possible names.

...

Named arguments to the worker constructor.

Returns

NULL (invisibly).

Method launch()

Launch all down workers in the crew.

Usage

class_crew$launch(tags = NULL)

Arguments

tags

Character vector of allowable tags of eligible workers.

Returns

NULL (invisibly).

Method sendable()

Determine if any worker is unassigned and ready to accept a new job.

Usage

class_crew$sendable(tags = NULL)

Arguments

tags

Character vector of allowable tags of eligible workers.

Returns

TRUE if the worker can accept a job and FALSE otherwise.

Method send()

Send a job to an available worker. Assumes sendable() on the crew returns TRUE.

Usage

class_crew$send(fun, args = list(), tags = NULL)

Arguments

fun

Function to run in the job.

args

Named list of function arguments to fun.

tags

Character vector of allowable tags of eligible workers.

Returns

NULL (invisibly).

Method receivable()

Determine if any worker in the crew is done with its current job and the job output is available for collection.

Usage

class_crew$receivable(tags = NULL)

Arguments

tags

Character vector of allowable tags of eligible workers.

Returns

TRUE if there exists an eligible worker with job output that can be accessed with receive(). FALSE otherwise.

Method receive()

Find a receivable worker, collect its job output, and free up the worker.

Usage

class_crew$receive(tags = NULL)

Arguments

tags

Character vector of allowable tags of eligible workers.

Details

Once collected and returned. the job output is deleted from the data store and no longer available to receive. receive() also marks the worker as "sendable" again (unblocked) which makes the worker ready for another job (send() method).

Returns

A named list of job output. The value element has the actual result of the job function, if successful. Other elements have job metadata such as the error message (if any), traceback, warnings, and runtime in seconds.

Method clear()

Clear worker input and output.

Usage

class_crew$clear(timeout = 60, wait = 1, down_only = TRUE, tags = NULL)

Arguments

timeout

Number of seconds to wait for the file deletion to succeed.

wait

Number of seconds to wait between iterations checking that the worker files were successfully removed from the data store.

down_only

Logical of length 1, whether to only clear down jobs.

tags

Character vector of allowable tags of eligible workers.

Details

Deletes worker input and output files.

Returns

NULL (invisibly)

Method shutdown()

Shut down one or more running workers.

Usage

class_crew$shutdown(
  workers = Inf,
  sendable_only = TRUE,
  up_only = TRUE,
  tags = NULL
)

Arguments

workers

Positive integer of length 1. Maximum number of workers to try to shut down. Does not count busy (non-sendable) workers.

sendable_only

Logical of length 1, whether to ignore workers that are not sendable.

up_only

Logical of length 1, whether to ignore workers that are not up.

tags

Character vector of allowable tags of eligible workers.

Details

This method loops through the workers from first to last in the worker list. If the worker is up (running) and sendable (able to receive a job) then the worker is sent a shutdown command. Otherwise, the loop moves on to another worker.

Returns

NULL (invisibly).

Method dismiss()

Delete one or more worker objects.

Usage

class_crew$dismiss(
  workers = Inf,
  sendable_only = TRUE,
  down_only = TRUE,
  tags = NULL
)

Arguments

workers

Positive integer of length 1. Maximum number of workers to try to dismiss. Does not count up or busy (non-sendable) workers (workers with unfinished jobs).

sendable_only

Logical of length 1, whether to ignore workers that are not sendable.

down_only

Logical of length 1, whether to ignore workers that are still up.

tags

Character vector of allowable tags of eligible workers.

Details

This method loops through the workers from first to last in the worker list. If the worker is down and sendable (able to receive a job) then the worker object is deleted from the crew. Otherwise, the loop moves on to another worker.

Returns

NULL (invisibly).

Method restart()

Restart one or more stuck workers.

Usage

class_crew$restart(workers = Inf, tags = NULL)

Arguments

workers

Positive integer of length 1. Maximum number of workers to try to restart. Does not count already unstuck workers.

tags

Character vector of allowable tags of eligible workers.

Details

This method loops through the workers from first to last in the worker list. If the worker is stuck, then the worker is restarted. Otherwise, the loop moves on to another worker.

Returns

NULL (invisibly).

Method validate()

Crew validator.

Usage

class_crew$validate()

Returns

NULL (invisibly).

Examples

crew <- class_crew$new()
crew$recruit(workers = 1, timeout = 1)
crew$send(fun = function(arg) paste("job", arg), args = list(arg = 1))
while (!crew$receivable()) Sys.sleep(0.1)
job <- crew$receive()
print(job$value)
#> [1] "job 1"
print(job$error)
#> NULL
crew$shutdown()