Worker class.

R6 class for a worker.

Details

A worker object is an interface to manage a single high-performance computing worker. Supported methods send jobs, receive output, and poll, etc. This particular worker class is the parent abstract class and is not usable on its own. Subclasses like class_worker_callr and class_worker_future are concrete and usable. In addition, worker needs a crew in order to be valid. So it is recommended to create a worker of a given subclass through the recruit() method of the crew. See the examples.

Public fields

name

Character of length 1, worker name.

crew

R6 crew object to which the worker belongs.

timeout

Positive numeric of length 1, number of seconds of idling for the worker to time out.

wait

Positive numeric of length 1, number of seconds in a polling interval, e.g. checking if an input job exists.

tags

Character vector of optional user-defined tags to select subsets of eligible workers for job submission and retrieval.

assigned

Logical of length 1, whether the worker already has a job to do.

Methods

Public methods

class_worker$new()

class_worker$tagged()

class_worker$sendable()

class_worker$send()

class_worker$receivable()

class_worker$receive()

class_worker$clear()

class_worker$shutdown()

class_worker$stuck()

class_worker$restart()

class_worker$validate()

Method new()

Worker constructor.

Usage

class_worker$new(
  name = basename(tempfile(pattern = "worker_")),
  crew = NULL,
  timeout = 60,
  wait = 0.1,
  tags = character(0)
)

Arguments

name

Character of length 1, worker name.

crew

R6 crew object to which the worker belongs.

timeout

Positive numeric of length 1, number of seconds that a worker can idle before timing out.

wait

Positive numeric of length 1, number of seconds in a polling interval, e.g. checking if an input job exists.

tags

Character vector of optional user-defined tags to mark eligible groups of workers in scheduling operations.

Returns

The new() method calls the constructor and returns a new worker object.

Method tagged()

Check if the worker has one or more of the tags in the argument.

Usage

class_worker$tagged(tags)

Arguments

tags

Character vector of tags to check.

Returns

TRUE if the worker has any of the tags in the tags argument. FALSE otherwise.

Method sendable()

Check if this worker is ready to accept a job.

Usage

class_worker$sendable()

Returns

TRUE if the worker is ready for a job (or not properly blocked) and FALSE otherwise.

Method send()

Send a job.

Usage

class_worker$send(fun = function() TRUE, args = list(), timeout = 60, wait = 1)

Arguments

fun

Function to run in the job. Should be completely self-contained in the body and arguments, without relying on the closure or global variables in the environment.

args

Named list of arguments to fun.

timeout

Number of seconds to wait for job data to send successfully.

wait

Number of seconds to wait between iterations checking if the job data was sent successfully.

Returns

NULL (invisibly).

Method receivable()

Check if job output is available to collect.

Usage

class_worker$receivable()

Returns

TRUE if a job output can be collected from the worker, FALSE otherwise.

Method receive()

Collect the results of a job and free up the worker.

Usage

class_worker$receive(timeout = 60, wait = 1)

Arguments

timeout

Number of seconds to wait for the file deletion to succeed.

wait

Number of seconds to wait between iterations checking that the worker files were successfully removed from the data store.

Details

Once collected and returned. the job output is deleted from the data store and no longer available to receive. receive() also marks the worker as "sendable" again (unblocked) which makes the worker ready for another job (send() method).

Returns

A named list of job output. The value element has the actual result of the job function, if successful. Other elements have job metadata such as the error message (if any), traceback, warnings, and runtime in seconds.

Method clear()

Clear worker input and output.

Usage

class_worker$clear(timeout = 60, wait = 1)

Arguments

timeout

Number of seconds to wait for the file deletion to succeed.

wait

Number of seconds to wait between iterations checking that the worker files were successfully removed from the data store.

Details

Deletes worker input and output files.

Returns

NULL (invisibly)

Method shutdown()

Gracefully shut down the worker.

Usage

class_worker$shutdown()

Details

The underlying worker process should promptly shut down if successful. A new send() or launch() call will re-launch the worker.

The default shutdown method sends a special shutdown job through the data store. This is not always reliable, e.g. if a worker freezes. Subclasses of class_worker should write their own shutdown methods that leverage the backend technology. to achieve more reliable shutdowns.

Returns

NULL (invisibly).

Method stuck()

Check if a worker is stuck.

Usage

class_worker$stuck()

Details

A worker is stuck if it is down, not sendable, and not receivable.

Returns

Logical of length 1, whether the worker is stuck.

Method restart()

Relaunch the worker if it is stuck.

Usage

class_worker$restart()

Returns

NULL (invisibly).

Method validate()

Worker validator.

Usage

class_worker$validate()

Returns

NULL (invisibly).

Examples

crew <- class_crew$new(worker_classes = list(class_worker_callr))
crew$recruit(workers = 1)
worker <- crew$workers[[1]]
worker$send(fun = function(arg) paste("job", arg), args = list(arg = 1))
while (!worker$receivable()) Sys.sleep(0.1)
job <- worker$receive()
print(job$value)
#> [1] "job 1"
print(job$error)
#> NULL
worker$shutdown()
processx::supervisor_kill()