Controller group class

R6 class for controller groups.

Details

See crew_controller_group().

See also

Public fields

controllers

List of R6 controller objects.

Methods

Public methods

crew_class_controller_group$new()

crew_class_controller_group$validate()

crew_class_controller_group$empty()

crew_class_controller_group$start()

crew_class_controller_group$launch()

crew_class_controller_group$scale()

crew_class_controller_group$push()

crew_class_controller_group$collect()

crew_class_controller_group$pop()

crew_class_controller_group$wait()

crew_class_controller_group$summary()

crew_class_controller_group$terminate()

Method new()

Multi-controller constructor.

Usage

crew_class_controller_group$new(controllers = NULL)

Arguments

controllers

List of R6 controller objects.

Returns

An R6 object with the controller group object.

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
persistent <- crew_controller_local(name = "persistent")
transient <- crew_controller_local(
  name = "transient",
  max_tasks = 1L
)
group <- crew_controller_group(persistent, transient)
group$connect()
group$push(name = "task", command = sqrt(4), controller = "transient")
group$wait()
group$pop()
group$controllers[["persistent"]]$launcher$running() # 0
group$controllers[["transient"]]$launcher$running() # 0
group$terminate()
crew_session_terminate()
}

Method validate()

Validate the router.

Usage

crew_class_controller_group$validate()

Returns

NULL (invisibly).

Method empty()

See if the controllers are empty.

Usage

crew_class_controller_group$empty(controllers = NULL)

Arguments

controllers

Character vector of controller names. Set to NULL to select all controllers.

Details

A controller is empty if it has no running tasks or completed tasks waiting to be retrieved with push().

Returns

TRUE if all the selected controllers are empty, FALSE otherwise.

Method start()

Start one or more controllers.

Usage

crew_class_controller_group$start(controllers = NULL)

Arguments

controllers

Character vector of controller names. Set to NULL to select all controllers.

Returns

NULL (invisibly).

Method launch()

Launch one or more workers on one or more controllers.

Usage

crew_class_controller_group$launch(n = 1L, controllers = NULL)

Arguments

n

Number of workers to launch in each controller selected.

controllers

Character vector of controller names. Set to NULL to select all controllers.

Returns

NULL (invisibly).

Method scale()

Automatically scale up the number of workers if needed in one or more controller objects.

Usage

crew_class_controller_group$scale(controllers = NULL)

Arguments

controllers

Character vector of controller names. Set to NULL to select all controllers.

Details

See the scale() method in individual controller classes.

Returns

NULL (invisibly).

Method push()

Push a task to the head of the task list.

Usage

crew_class_controller_group$push(
  command,
  data = list(),
  globals = list(),
  substitute = TRUE,
  seed = sample.int(n = 1000000000L, size = 1L),
  packages = character(0),
  library = NULL,
  seconds_timeout = NULL,
  scale = TRUE,
  name = NULL,
  controller = NULL
)

Arguments

command

Language object with R code to run.

data

Named list of local data objects in the evaluation environment.

globals

Named list of objects to temporarily assign to the global environment for the task. See the reset_globals argument of crew_controller_local().

substitute

Logical of length 1, whether to call base::substitute() on the supplied value of the command argument. If TRUE (default) then command is quoted literally as you write it, e.g. push(command = your_function_call()). If FALSE, then crew assumes command is a language object and you are passing its value, e.g. push(command = quote(your_function_call())). substitute = TRUE is appropriate for interactive use, whereas substitute = FALSE is meant for automated R programs that invoke crew controllers.

seed

Integer of length 1 with the pseudo-random number generator seed to temporarily set for the evaluation of the task. At the end of the task, the seed is restored.

packages

Character vector of packages to load for the task.

library

Library path to load the packages. See the lib.loc argument of require().

seconds_timeout

Optional task timeout passed to the .timeout argument of mirai::mirai() (after converting to milliseconds).

scale

Logical, whether to automatically scale workers to meet demand. If TRUE, then collect() runs first so demand can be properly assessed before scaling and the number of workers is not too high.

name

Optional name of the task. Replaced with a random name if NULL or in conflict with an existing name in the task list.

controller

Character of length 1, name of the controller to submit the task. If NULL, the controller defaults to the first controller in the list.

Returns

NULL (invisibly).

Method collect()

Check for done tasks and move the results to the results list.

Usage

crew_class_controller_group$collect(controllers = NULL)

Arguments

controllers

Character vector of controller names. Set to NULL to select all controllers.

Returns

NULL (invisibly). Removes elements from the queue list as applicable and moves them to the results list.

Method pop()

Pop a completed task from the results data frame.

Usage

crew_class_controller_group$pop(scale = TRUE, controllers = NULL)

Arguments

scale

Logical, whether to automatically scale workers to meet demand. If TRUE, then collect() runs first so demand can be properly assessed before scaling and the number of workers is not too high. Scaling up on pop() may be important for transient or nearly transient workers that tend to drop off quickly after doing little work.

controllers

Character vector of controller names. Set to NULL to select all controllers.

Returns

If there is a completed task available to collect, the return value is a one-row data frame with the results, warnings, and errors. Otherwise, if there are no results available to collect, the return value is NULL.

Method wait()

Wait for tasks.

Usage

crew_class_controller_group$wait(
  mode = "all",
  seconds_interval = 0.001,
  seconds_timeout = Inf,
  controllers = NULL
)

Arguments

mode

If mode is "all", then the method waits for all tasks to complete. If mode is "one", then it waits until a one task is complete.

seconds_interval

Number of seconds to wait between polling intervals while checking for results.

seconds_timeout

Timeout length in seconds waiting for results to become available.

controllers

Character vector of controller names. Set to NULL to select all controllers.

Details

The wait() method blocks the calling R session, repeatedly auto-scales workers for tasks that need them, and repeatedly collects results. The function runs until it either times out or one of the controllers reaches the stopping condition based on the mode argument.

Returns

NULL (invisibly). Call pop() to get the result.

Method summary()

Summarize the workers of one or more controllers.

Usage

crew_class_controller_group$summary(
  columns = tidyselect::everything(),
  controllers = NULL
)

Arguments

columns

Tidyselect expression to select a subset of columns. Examples include columns = contains("worker") and columns = starts_with("tasks").

controllers

Character vector of controller names. Set to NULL to select all controllers.

Returns

A data frame of aggregated worker summary statistics of all the selected controllers. It has one row per worker websocket, and the rows are grouped by controller. See the documentation of the summary() method of the controller class for specific information about the columns in the output.

Method terminate()

Terminate the workers and disconnect the router for one or more controllers.

Usage

crew_class_controller_group$terminate(controllers = NULL)

Arguments

controllers

Character vector of controller names. Set to NULL to select all controllers.

Returns

NULL (invisibly).

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
persistent <- crew_controller_local(name = "persistent")
transient <- crew_controller_local(
  name = "transient",
  max_tasks = 1L
)
group <- crew_controller_group(persistent, transient)
group$connect()
group$push(name = "task", command = sqrt(4), controller = "transient")
group$wait()
group$pop()
group$controllers[["persistent"]]$launcher$running() # 0
group$controllers[["transient"]]$launcher$running() # 0
group$terminate()
crew_session_terminate()
}

## ------------------------------------------------
## Method `crew_class_controller_group$new`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
persistent <- crew_controller_local(name = "persistent")
transient <- crew_controller_local(
  name = "transient",
  max_tasks = 1L
)
group <- crew_controller_group(persistent, transient)
group$connect()
group$push(name = "task", command = sqrt(4), controller = "transient")
group$wait()
group$pop()
group$controllers[["persistent"]]$launcher$running() # 0
group$controllers[["transient"]]$launcher$running() # 0
group$terminate()
crew_session_terminate()
}