Launcher abstract class

R6 abstract class to build other subclasses which launch and manage workers.

See also

Public fields

workers

Data frame of worker information.

name

Name of the launcher.

seconds_launch

See crew_launcher().

seconds_interval

See crew_launcher().

seconds_timeout

See crew_launcher().

seconds_idle

See crew_launcher().

seconds_wall

See crew_launcher().

seconds_exit

See crew_launcher().

tasks_max

See crew_launcher().

tasks_timers

See crew_launcher().

reset_globals

See crew_launcher().

reset_packages

See crew_launcher().

reset_options

See crew_launcher().

garbage_collection

See crew_launcher().

Methods

Public methods

crew_class_launcher$new()

crew_class_launcher$validate()

crew_class_launcher$settings()

crew_class_launcher$call()

crew_class_launcher$start()

crew_class_launcher$launch()

crew_class_launcher$launching()

crew_class_launcher$terminate()

crew_class_launcher$terminate_worker()

Method new()

Launcher constructor.

Usage

crew_class_launcher$new(
  name = NULL,
  seconds_launch = NULL,
  seconds_interval = NULL,
  seconds_timeout = NULL,
  seconds_idle = NULL,
  seconds_wall = NULL,
  seconds_exit = NULL,
  tasks_max = NULL,
  tasks_timers = NULL,
  reset_globals = NULL,
  reset_packages = NULL,
  reset_options = NULL,
  garbage_collection = NULL
)

Arguments

name

See crew_launcher().

seconds_launch

See crew_launcher().

seconds_interval

See crew_launcher().

seconds_timeout

See crew_launcher().

seconds_idle

See crew_launcher(). See the idletime argument of mirai::server().

seconds_wall

See crew_launcher().

seconds_exit

See crew_launcher().

tasks_max

See crew_launcher().

tasks_timers

See crew_launcher().

reset_globals

See crew_launcher().

reset_packages

See crew_launcher().

reset_options

See crew_launcher().

garbage_collection

See crew_launcher().

Returns

An R6 object with the launcher.

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
router <- crew_router()
router$listen()
launcher <- crew_launcher_local()
launcher$populate(sockets = router$sockets)
launcher$launch()
m <- mirai::mirai("result", .compute = router$name)
Sys.sleep(0.25)
m$data
router$terminate()
crew_session_terminate()
}

Method validate()

Validate the launcher.

Usage

crew_class_launcher$validate()

Returns

NULL (invisibly).

Method settings()

List of arguments for mirai::server().

Usage

crew_class_launcher$settings(socket)

Arguments

socket

Character of length 1, websocket address of the worker to launch.

Returns

List of arguments for mirai::server().

Method call()

Create a call to crew_worker() to help create custom launchers.

Usage

crew_class_launcher$call(socket, launcher, worker, instance)

Arguments

socket

Socket where the worker will receive tasks.

launcher

Character of length 1, name of the launcher.

worker

Positive integer of length 1, index of the worker. This worker index remains the same even when the current instance of the worker exits and a new instance launches.

instance

Character of length 1 to uniquely identify the instance of the worker.

Returns

Character of length 1 with a call to crew_worker().

Examples

launcher <- crew_launcher_local()
launcher$call(
  socket = "ws://127.0.0.1:5000/3/cba033e58",
  launcher = "launcher_a",
  worker = 3L,
  instance = "cba033e58"
)

Method start()

Start the launcher.

Usage

crew_class_launcher$start(workers = 1L)

Arguments

workers

Positive integer of length 1, number of workers to allow.

Details

Creates the workers data frame. Meant to be called once at the beginning of the launcher life cycle.

Returns

NULL (invisibly).

Method launch()

Launch a worker.

Usage

crew_class_launcher$launch(index, socket = NULL)

Arguments

index

Positive integer of length 1, index of the worker to launch.

socket

Character of length 1, sockets of the worker to launch.

Returns

NULL (invisibly).

Method launching()

Show whether each worker is launching.

Usage

crew_class_launcher$launching()

Details

A worker is considered "launching" if it was started recently (seconds_launch seconds ago or sooner).

Returns

A logical vector indicating which workers are launching.

Method terminate()

Terminate one or more workers.

Usage

crew_class_launcher$terminate(indexes = NULL)

Arguments

indexes

Integer vector of the indexes of the workers to terminate. If NULL, all current workers are terminated.

Returns

NULL (invisibly).

Method terminate_worker()

Abstract method.

Usage

crew_class_launcher$terminate_worker(handle)

Arguments

handle

A handle object previously returned by launch_worker() which allows the termination of the worker.

Details

Does not actually terminate a worker. This method is a placeholder, and its presence allows manual worker termination to be optional.

Returns

NULL (invisibly).

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
router <- crew_router()
router$listen()
launcher <- crew_launcher_local()
launcher$populate(sockets = router$sockets)
launcher$launch()
m <- mirai::mirai("result", .compute = router$name)
Sys.sleep(0.25)
m$data
router$terminate()
crew_session_terminate()
}

## ------------------------------------------------
## Method `crew_class_launcher$new`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
router <- crew_router()
router$listen()
launcher <- crew_launcher_local()
launcher$populate(sockets = router$sockets)
launcher$launch()
m <- mirai::mirai("result", .compute = router$name)
Sys.sleep(0.25)
m$data
router$terminate()
crew_session_terminate()
}

## ------------------------------------------------
## Method `crew_class_launcher$call`
## ------------------------------------------------

launcher <- crew_launcher_local()
launcher$call(
  socket = "ws://127.0.0.1:5000/3/cba033e58",
  launcher = "launcher_a",
  worker = 3L,
  instance = "cba033e58"
)
#> [1] "crew::crew_worker(settings = list(url = \"ws://127.0.0.1:5000/3/cba033e58\", maxtasks = Inf, idletime = Inf, walltime = Inf, timerstart = 0L, exitlinger = 1000, cleanup = 1L, asyncdial = FALSE), launcher = \"launcher_a\", worker = 3L, instance = \"cba033e58\")"