Create a controller.

Create an R6 object to submit tasks and launch workers.

crew_controller(router, launcher, auto_scale = "demand")

Arguments

router

An R6 router object created by crew_router().

launcher

An R6 launcher object created by one of the crew_launcher_*() functions such as crew_launcher_local().

auto_scale

Character of length 1, name of the method for automatically scaling workers to meet demand. NULL to default to "demand". Possible values include the following:

"demand": just after pushing a new task in push(), launch min(n, max(0, t - w)) workers, where n is the maximum number of workers, t is the number of queued tasks, and w is the current number of workers already running. In other words, scale up the number of workers to meet the current demand. If you trust tasks not to crash workers, this is a good choice. But if you think a task may always crash a worker (e.g. segmentation fault or maxed out memory) then this could be somewhat risky because mirai resubmits failed tasks behind the scenes and crew responds by re-launching workers. If you are worried about this scenario, choose auto_scale = "one" instead, which will only launch up to one worker whenever a task is pushed.

"one": just after pushing a new task in push(), launch a one worker if demand min(n, max(0, t - w)) is greater than 0.

"none": do not auto-scale at all.

See also

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
crew_session_start()
router <- crew_router()
launcher <- crew_launcher_local()
controller <- crew_controller(router = router, launcher = launcher)
controller$start()
controller$push(name = "task", command = sqrt(4))
controller$wait()
controller$pop()
controller$terminate()
crew_session_terminate()
}