Evaluate an R command and return results as a monad.

Not a user-side function. Do not call directly.

crew_eval(
  command,
  data = list(),
  globals = list(),
  seed = sample.int(n = 1000000000L, size = 1L),
  garbage_collection = FALSE,
  packages = character(0),
  library = NULL
)

Arguments

command

Language object with R code to run.

data

Named list of local data objects in the evaluation environment.

globals

Named list of objects to temporarily assign to the global environment for the task. At the end of the task, these values are reset to their previous values.

seed

Integer of length 1 with the pseudo-random number generator seed to temporarily set for the evaluation of the task. At the end of the task, the seed is restored.

garbage_collection

Logical, whether to run garbage collection with gc() before running the task.

packages

Character vector of packages to load for the task.

library

Library path to load the packages. See the lib.loc argument of require().

Value

A monad object with results and metadata.

Details

The crew_eval() function evaluates an R expression in an encapsulated environment and returns a monad with the results, including warnings and error messages if applicable. The random number generator seed, globals, and global options are restored to their original values on exit.

See also

Examples

crew_eval(quote(1 + 1))
#> # A tibble: 1 × 10
#>   name  command result    seconds     seed error trace…¹ warni…² socke…³ socke…⁴
#>   <chr> <chr>   <list>      <dbl>    <int> <chr> <chr>   <chr>   <chr>   <chr>  
#> 1 NA    1 + 1   <dbl [1]>       0   7.24e8 NA    NA      NA      NA      NA     
#> # … with abbreviated variable names ¹​traceback, ²​warnings, ³​socket_data,
#> #   ⁴​socket_session